style: rename package command to processor
