+++ node/src/main/java/com/github/bucketonhead/service/processor/CommandProcessor.java
@@ -1,4 +1,4 @@
-package com.github.bucketonhead.service.command;
+package com.github.bucketonhead.service.processor;
 
 import com.github.bucketonhead.entity.AppUser;
 import org.telegram.telegrambots.meta.api.objects.Message;
