+++ node/src/main/java/com/github/bucketonhead/service/processor/basic/BasicService.java
@@ -1,7 +1,7 @@
-package com.github.bucketonhead.service.command.basic;
+package com.github.bucketonhead.service.processor.basic;
 
 import com.github.bucketonhead.entity.AppUser;
-import com.github.bucketonhead.service.command.CommandProcessor;
+import com.github.bucketonhead.service.processor.CommandProcessor;
 
 public interface BasicService extends CommandProcessor {
     String processCancelCommand(AppUser user);
