+++ node/src/main/java/com/github/bucketonhead/service/processor/main/MainService.java
@@ -1,4 +1,4 @@
-package com.github.bucketonhead.service.command.main;
+package com.github.bucketonhead.service.processor.main;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
