+++ node/src/main/java/com/github/bucketonhead/service/processor/task/TaskService.java
@@ -1,7 +1,7 @@
-package com.github.bucketonhead.service.command.task;
+package com.github.bucketonhead.service.processor.task;
 
 import com.github.bucketonhead.entity.AppUser;
-import com.github.bucketonhead.service.command.CommandProcessor;
+import com.github.bucketonhead.service.processor.CommandProcessor;
 import org.telegram.telegrambots.meta.api.objects.Message;
 
 public interface TaskService extends CommandProcessor {
