+++ node/src/main/java/com/github/bucketonhead/service/rabbitmq/impl/ConsumerServiceImpl.java
@@ -1,7 +1,7 @@
 package com.github.bucketonhead.service.rabbitmq.impl;
 
 import com.github.bucketonhead.constants.RabbitQueue;
-import com.github.bucketonhead.service.command.main.MainService;
+import com.github.bucketonhead.service.processor.main.MainService;
 import com.github.bucketonhead.service.rabbitmq.ConsumerService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
